IB/mad: Minor fix when updating slid in struct mad_rmpp_recv
Commit 49745a237721 ("IB/mad: Change slid in RMPP recv from 16 to 32
bits.") increased the length of the slid field in struct mad_rmpp_recv
from 16 to 32 bits. There was still a buggy typecast that
limited this to 16 bits. This patch fixes it.

+++ drivers/infiniband/core/mad_rmpp.c
@@ -317,7 +317,7 @@ create_rmpp_recv(struct ib_mad_agent_private *agent,
 	mad_hdr = &mad_recv_wc->recv_buf.mad->mad_hdr;
 	rmpp_recv->tid = mad_hdr->tid;
 	rmpp_recv->src_qp = mad_recv_wc->wc->src_qp;
-	rmpp_recv->slid = (u16)mad_recv_wc->wc->slid;
+	rmpp_recv->slid = mad_recv_wc->wc->slid;
 	rmpp_recv->mgmt_class = mad_hdr->mgmt_class;
 	rmpp_recv->class_version = mad_hdr->class_version;
 	rmpp_recv->method  = mad_hdr->method;
